Overhaul the parser functions and update their docs
