+++ .dockerignore
@@ -1,4 +1,3 @@
-clowder/
 inputs/
 outputs/
 venv/
