+++ pipeline/getChallengerIndexSpecies.py
